project/requirements.txt
numpy
Pillow
environs